backend/requirements.txt
astapi
uvicorn[standard] 
sqlalchemy
psycopg[binary]
pgvector